Add event loop
