+++ node/src/client/mod.rs
@@ -133,6 +133,8 @@ impl<N: Network, C: ConsensusStorage<N>> Client<N, C> {
         node.initialize_routing().await;
         // Initialize the sync module.
         node.initialize_sync();
+        // Initialize the notification message loop.
+        node.handles.lock().push(crate::start_notification_message_loop());
         // Pass the node to the signal handler.
         let _ = signal_node.set(node.clone());
         // Return the node.
